[feat]근태 관리자 코드레빗(#43)

+++ attendance/src/main/java/com/reverse/attendance/dto/response/PayrollAttendanceResponse.java
@@ -1,5 +1,6 @@
 package com.reverse.attendance.dto.response;
 
+import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -17,12 +18,12 @@ public class PayrollAttendanceResponse {
 
     // 1. 기본급 계산용
     private int normalWorkDays; // 정상 근무 일수
-    private double totalNormalWorkHours; // 정상 근무 총 시간
+    private BigDecimal totalNormalWorkHours; // 정상 근무 총 시간
 
     // 2. 추가 수당 계산용
-    private double totalOvertimeHours; // 연장근무 총 시간 (1.5배 수당)
-    private double nightWorkHours; // 야간근무 총 시간
-    private double holidayWorkHours; // 휴일근무 총 시간
+    private BigDecimal totalOvertimeHours; // 연장근무 총 시간
+    private BigDecimal nightWorkHours; // 야간근무 총 시간
+    private BigDecimal holidayWorkHours; // 휴일근무 총 시간
     private int businessTripDays; // 출장 일수
 
     // 3. 급여 차감용
@@ -31,6 +32,6 @@ public class PayrollAttendanceResponse {
     private int absentDays; // 무단 결근 일수
 
     // 4. 휴가 관련
-    private double paidLeaveDays; // 유급 휴가 사용 일수
-    private double unpaidLeaveDays; // 무급 휴가 사용 일수
+    private BigDecimal paidLeaveDays; // 유급 휴가 사용 일수
+    private BigDecimal unpaidLeaveDays; // 무급 휴가 사용 일수
 }
